[DSR - Core] Change attribute 'person_id' to std::int32_t

+++ libs/dsr/core/include/dsr/core/types/type_checking/dsr_attr_name.h
@@ -128,7 +128,7 @@ REGISTER_TYPE(person_personal_y_pos, std::reference_wrapper<const std::vector<fl
 REGISTER_TYPE(person_sharedWidth, std::reference_wrapper<const std::vector<float>>,true)
 REGISTER_TYPE(person_intimate_x_pos, std::reference_wrapper<const std::vector<float>>,true)
 REGISTER_TYPE(person_intimate_y_pos, std::reference_wrapper<const std::vector<float>>,true)
-REGISTER_TYPE(person_id, std::uint64_t, false)
+REGISTER_TYPE(person_id, std::int32_t, false)
 REGISTER_TYPE(distance_to_robot, float, false)
 REGISTER_TYPE(lambda_cont, std::int32_t, false)
 REGISTER_TYPE(is_ready, bool, false)
